requirements.txt
Django>=4.1,<5.0
djangorestframework>=3.13,<4.0
djongo>=1.3.6
pymongo>=3.12,<4.0